pyproject.toml
[project]
name = "travel-planner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "black>=25.1.0",
    "flake8>=7.3.0",
    "isort>=6.0.1",
    "langchain-openai>=0.3.32",
    "langchain-tavily>=0.2.11",
    "langgraph>=0.6.7",
    "pre-commit>=4.3.0",
    "python-dotenv>=1.1.1",
]
